+++ lib/src/config.dart
@@ -24,9 +24,11 @@ class Config {
       this.enableSkinTones = true,
       this.showRecentsTab = true,
       this.recentsLimit = 28,
-      this.noRecentsText = 'No Recents',
-      this.noRecentsStyle =
-          const TextStyle(fontSize: 20, color: Colors.black26),
+      this.noRecentsText = const Text(
+        'No Recents',
+        style: TextStyle(fontSize: 20, color: Colors.black26),
+        textAlign: TextAlign.center,
+      ),
       this.tabIndicatorAnimDuration = kTabScrollDuration,
       this.categoryIcons = const CategoryIcons(),
       this.buttonMode = ButtonMode.MATERIAL});
@@ -81,11 +83,8 @@ class Config {
   /// Limit of recently used emoji that will be saved
   final int recentsLimit;
 
-  /// The text to be displayed if no recent emojis to display
-  final String noRecentsText;
-
-  /// The text style for [noRecentsText]
-  final TextStyle noRecentsStyle;
+  /// A widget (usually [Text]) to be displayed if no recent emojis to display
+  final Widget noRecentsText;
 
   /// Duration of tab indicator to animate to next category
   final Duration tabIndicatorAnimDuration;
@@ -148,7 +147,6 @@ class Config {
         other.showRecentsTab == showRecentsTab &&
         other.recentsLimit == recentsLimit &&
         other.noRecentsText == noRecentsText &&
-        other.noRecentsStyle == noRecentsStyle &&
         other.tabIndicatorAnimDuration == tabIndicatorAnimDuration &&
         other.categoryIcons == categoryIcons &&
         other.buttonMode == buttonMode;
@@ -173,7 +171,6 @@ class Config {
       showRecentsTab.hashCode ^
       recentsLimit.hashCode ^
       noRecentsText.hashCode ^
-      noRecentsStyle.hashCode ^
       tabIndicatorAnimDuration.hashCode ^
       categoryIcons.hashCode ^
       buttonMode.hashCode;